fix: incorrect type for v fixes https://github.com/rarible/protocol-issues/issues/107 (#90)

+++ gateway/src/test/kotlin/com/rarible/protocol/gateway/route/ProtocolRouteLocatorTest.kt
@@ -40,7 +40,7 @@ internal class ProtocolRouteLocatorTest : AbstractIntegrationTest() {
             signature = NftSignatureDto(
                 r = Binary.apply("0x12"),
                 s = Binary.apply("0x13"),
-                v = Byte.MIN_VALUE
+                v = Byte.MIN_VALUE.toInt()
             )
         )
         mockNftServerClient
